Refactor Timings for @aikar, fixes #1369
