+++ src/main/java/cn/nukkit/command/Command.java
@@ -9,9 +9,9 @@ import cn.nukkit.command.data.CommandParameter;
 import cn.nukkit.lang.TextContainer;
 import cn.nukkit.lang.TranslationContainer;
 import cn.nukkit.permission.Permissible;
-import cn.nukkit.timings.Timing;
-import cn.nukkit.timings.Timings;
 import cn.nukkit.utils.TextFormat;
+import co.aikar.timings.Timing;
+import co.aikar.timings.Timings;
 
 import java.util.HashMap;
 import java.util.Map;
