+++ src/main/java/cn/nukkit/command/defaults/TimingsCommand.java
@@ -3,8 +3,8 @@ package cn.nukkit.command.defaults;
 import cn.nukkit.command.CommandSender;
 import cn.nukkit.command.data.CommandParameter;
 import cn.nukkit.lang.TranslationContainer;
-import cn.nukkit.timings.Timings;
-import cn.nukkit.timings.TimingsExport;
+import co.aikar.timings.Timings;
+import co.aikar.timings.TimingsExport;
 
 /**
  * @author fromgate
