+++ src/main/java/cn/nukkit/entity/Entity.java
@@ -31,10 +31,10 @@ import cn.nukkit.network.protocol.SetEntityDataPacket;
 import cn.nukkit.network.protocol.SetEntityMotionPacket;
 import cn.nukkit.plugin.Plugin;
 import cn.nukkit.potion.Effect;
-import cn.nukkit.timings.Timing;
-import cn.nukkit.timings.Timings;
-import cn.nukkit.timings.TimingsHistory;
 import cn.nukkit.utils.ChunkException;
+import co.aikar.timings.Timing;
+import co.aikar.timings.Timings;
+import co.aikar.timings.TimingsHistory;
 
 import java.lang.reflect.Constructor;
 import java.util.*;
