+++ src/main/java/co/aikar/timings/TimingsHistoryEntry.java
@@ -1,0 +1,47 @@
+/*
+ * This file is licensed under the MIT License (MIT).
+ *
+ * Copyright (c) 2014 Daniel Ennis <http://aikar.co>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package co.aikar.timings;
+
+import com.google.gson.JsonArray;
+
+class TimingsHistoryEntry {
+    final TimingData data;
+    final TimingData[] children;
+
+    TimingsHistoryEntry(Timing timing) {
+        this.data = timing.record.clone();
+        this.children = new TimingData[timing.children.size()];
+
+        int i = 0;
+        for (TimingData child : timing.children.values()) {
+            this.children[i++] = child.clone();
+        }
+    }
+
+    JsonArray export() {
+        JsonArray json = this.data.export();
+        if (this.children.length > 0) json.add(JsonUtil.mapToArray(this.children, TimingData::export));
+        return json;
+    }
+}
